+++ src/main/java/com/ggukmoney/beanzip/domain/ranking/entity/RankingType.java
@@ -1,5 +1,6 @@
 package com.ggukmoney.beanzip.domain.ranking.entity;
 
 public enum RankingType {
-    ALL_TIME
+    ALL_TIME,
+    WEEKLY
 }